move classes around
