+++ packages/Ecotone/src/Modelling/AggregateFlow/LoadAggregate/LoadAggregateServiceBuilder.php
@@ -12,8 +12,8 @@ use Ecotone\Messaging\Handler\Enricher\PropertyReaderAccessor;
 use Ecotone\Messaging\Handler\ExpressionEvaluationService;
 use Ecotone\Messaging\Handler\Processor\InterceptedMessageProcessorBuilder;
 use Ecotone\Messaging\Handler\TypeDescriptor;
-use Ecotone\Modelling\AggregateFlow\AllAggregateRepository;
 use Ecotone\Modelling\Attribute\TargetAggregateVersion;
+use Ecotone\Modelling\Repository\AllAggregateRepository;
 
 /**
  * licence Apache-2.0
