+++ packages/Ecotone/src/Modelling/Config/AggregrateModule.php
@@ -36,7 +36,6 @@ use Ecotone\Messaging\Handler\Transformer\TransformerProcessorBuilder;
 use Ecotone\Messaging\Handler\TypeDescriptor;
 use Ecotone\Messaging\MessageConverter\DefaultHeaderMapper;
 use Ecotone\Messaging\Support\Assert;
-use Ecotone\Modelling\AggregateFlow\AggregateRepositoryBuilder;
 use Ecotone\Modelling\AggregateFlow\CallAggregate\CallAggregateServiceBuilder;
 use Ecotone\Modelling\AggregateFlow\LoadAggregate\LoadAggregateMode;
 use Ecotone\Modelling\AggregateFlow\LoadAggregate\LoadAggregateServiceBuilder;
@@ -45,7 +44,6 @@ use Ecotone\Modelling\AggregateFlow\SaveAggregate\AggregateResolver\AggregateDef
 use Ecotone\Modelling\AggregateFlow\SaveAggregate\AggregateResolver\AggregateDefinitionResolver;
 use Ecotone\Modelling\AggregateFlow\SaveAggregate\AggregateResolver\AggregateResolver;
 use Ecotone\Modelling\AggregateFlow\SaveAggregate\SaveAggregateServiceBuilder;
-use Ecotone\Modelling\AggregateFlow\StandardRepositoryAdapterBuilder;
 use Ecotone\Modelling\AggregateIdentifierRetrevingServiceBuilder;
 use Ecotone\Modelling\AggregateMessage;
 use Ecotone\Modelling\Attribute\Aggregate;
@@ -60,6 +58,8 @@ use Ecotone\Modelling\EventSourcingExecutor\EventSourcingHandlerExecutorBuilder;
 use Ecotone\Modelling\EventSourcingExecutor\GroupedEventSourcingExecutor;
 use Ecotone\Modelling\EventSourcingExecutor\OpenCoreAggregateMethodInvoker;
 use Ecotone\Modelling\FetchAggregate;
+use Ecotone\Modelling\Repository\AggregateRepositoryBuilder;
+use Ecotone\Modelling\Repository\StandardRepositoryAdapterBuilder;
 use Ecotone\Modelling\RepositoryBuilder;
 use InvalidArgumentException;
 use Ramsey\Uuid\Uuid;
