+++ packages/Ecotone/src/Modelling/Config/EventSourcedRepositoryModule.php
@@ -7,6 +7,7 @@ declare(strict_types=1);
 namespace Ecotone\Modelling\Config;
 
 use Ecotone\AnnotationFinder\AnnotationFinder;
+use Ecotone\EventSourcing\EventSourcedRepositoryAdapterBuilder;
 use Ecotone\Messaging\Attribute\ModuleAnnotation;
 use Ecotone\Messaging\Config\Annotation\AnnotationModule;
 use Ecotone\Messaging\Config\Configuration;
@@ -14,7 +15,6 @@ use Ecotone\Messaging\Config\ModulePackageList;
 use Ecotone\Messaging\Config\ModuleReferenceSearchService;
 use Ecotone\Messaging\Config\ServiceConfiguration;
 use Ecotone\Messaging\Handler\InterfaceToCallRegistry;
-use Ecotone\Modelling\AggregateFlow\EventSourcedRepositoryAdapterBuilder;
 use Ecotone\Modelling\BaseEventSourcingConfiguration;
 
 #[ModuleAnnotation]
